(ol-concourse-lib) fix: Properly tag and retrieve Learn AI docker images for RC/Production
The git tag does not get set on the release-candidate branch, so it is necessary to
retrieve the version details from the code for tagging the image. This udpates the RC
git resource to not filter on tags that don't exist yet, and instead fetch the version
via bash command. Also update the registry-image resource to properly filter on the
tagged versions to propagate to the Pulumi code